+++ src/api/consumer/lib/build.ts
@@ -5,8 +5,13 @@ import { COMPONENT_ORIGINS } from '../../../constants';
 import loader from '../../../cli/loader';
 import { BEFORE_LOADING_COMPONENTS } from '../../../cli/loader/loader-messages';
 
-export async function build(id: string, noCache: boolean, verbose: boolean): Promise<string[] | undefined> {
-  const consumer = await loadConsumer();
+export async function build(
+  id: string,
+  noCache: boolean,
+  verbose: boolean,
+  workspaceDir = process.cwd()
+): Promise<string[] | undefined> {
+  const consumer = await loadConsumer(workspaceDir);
   const bitId = consumer.getParsedId(id);
   const component: Component = await consumer.loadComponent(bitId);
   const results = await consumer.scope.buildMultiple([component], consumer, noCache, verbose);